% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlsR-package.R
\docType{package}
\name{tlsR-package}
\alias{tlsR-package}
\alias{tlsR}
\title{tlsR: Detection and Spatial Analysis of Tertiary Lymphoid Structures}
\description{
Fast, reproducible detection and quantitative analysis of tertiary lymphoid
structures (TLS) in multiplexed tissue imaging data.
}
\section{Typical workflow}{

\enumerate{
  \item Load or prepare a named list of data frames (\code{ldata}), one per
        tissue sample. Each data frame must contain columns \code{x}, \code{y}
        (spatial coordinates in microns), and \code{phenotype} (character:
        \code{"B cell"} / \code{"T cell"} / other).
  \item Run \code{\link{detect_TLS}} to label B+T co-localised regions.
  \item (Optional) Run \code{\link{scan_clustering}} to identify windows of
        significant immune clustering via local Ripley's L.
  \item Run \code{\link{calc_icat}} to score the internal linearity/organisation
        of each detected TLS.
  \item Run \code{\link{detect_tic}} to identify T-cell clusters outside TLS.
  \item Use \code{\link{summarize_TLS}} to obtain a tidy summary table.
  \item Use \code{\link{plot_TLS}} to produce publication-ready spatial plots.
}
}

\references{
Amiryousefi et al. (2025) \doi{10.1101/2025.09.21.677465}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/labsyspharm/tlsR}
}

}
\author{
\strong{Maintainer}: Ali Amiryousefi \email{ali_amiryousefi@hms.harvard.edu} (\href{https://orcid.org/0000-0002-6317-3860}{ORCID})

Authors:
\itemize{
  \item Jeremiah Wala \email{jeremiah_wala@dfci.harvard.edu} (\href{https://orcid.org/0000-0001-6591-1620}{ORCID})
}

Other contributors:
\itemize{
  \item Peter Sorger (\href{https://orcid.org/0000-0002-3364-1838}{ORCID}) [contributor]
}

}
